refactor: move FilterItems into components folder
