+++ src/components/QuickFilter/QuickFilter.tsx
@@ -5,7 +5,7 @@ import MapObjectsButton from '@/ui/buttons/MapObjectsButton'
 import CatalogViewButton from '@/components/QuickFilter/CatalogViewButton'
 import { flatPlural } from '@/features/utility/pluralRules'
 import { QuickFilters } from '@/types/FiltersType'
-import { FilterItems } from '@/features/FilterItems'
+import { FilterItems } from '@/components/FilterItems'
 import SortButton from '@/components/QuickFilter/SortButton'
 import ResetFiltersButton from '@/components/QuickFilter/ResetFiltersButton'
 import { FiltersTags } from '@/components/QuickFilter/FilterTags'
