+++ src/ui/popups/FilterPopup/FilterPopup.tsx
@@ -6,7 +6,7 @@ import Button from '@/ui/buttons/Button'
 import { getFilters } from '@/globals/api/methods/getFilters'
 import { FilterBlock } from '@/types/FiltersType'
 import { IsDesktop, IsMobile } from '@/features/utility/adaptive'
-import { FilterItems } from '@/features/FilterItems'
+import { FilterItems } from '@/components/FilterItems'
 import ResetFiltersButton from '@/components/QuickFilter/ResetFiltersButton'
 import { FiltersTags } from '@/components/QuickFilter/FilterTags'
 import FilterBlockWrapper from '@/ui/popups/FilterPopup/FilterBlockWrapper'
